HUD Updates
Removed wealth count, added a clip meter instead. Separate panel above
ammo for powers, as sort of "badge" icons that could be sprite-swapped
as needed.

Inserted black image for transition and made two animation clips ready
for use in fading into or out of a level.

Added panels for level name and team name as well, for aesthetics